Split tests one per addon
